Change size_t to int in for_trivial_lt

+++ examples/intraprocedural/control/for_trivial_lt.cpp
@@ -3,7 +3,7 @@
 int main()
 {
     int * arr = new int[10];
-    for (size_t x = 0; x < 10; ++x)
+    for (int x = 0; x < 10; ++x)
         arr[x] = 5;
 
     return 0;
